pyproject.toml
[tool.poetry]
name = "notetake"
version = "0.1.0"
description = "Personal note taking app"
authors = ["Mike Vink <mike1994vink@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
python-lsp-server = {extras = ["all"], version = "^1.2.1"}
typer = "^0.3.2"
pandoc = {git = "https://github.com/boisgera/pandoc.git"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cli = "notetake.cli:app"
